Fix up the main tox ebuild
This removes the capability to link against nacl, as linking against
nacl disables all dynamic linking, which can't work for any client
except toxic. Also adds use flags for nTox and the dht daemon, to build
them on demand. An initscript for the daemon is included.

Also fix repoman warnings in net-libs/tox